requirements.txt
alembic==1.17.1
fastapi==0.121.0
passlib==1.7.4
pydantic==2.12.4
pydantic_settings==2.11.0
PyJWT==2.10.1
SQLAlchemy==2.0.44
